# Conflicts:
#	app/src/main/java/org/t_robop/y_ogawara/tancolle/MainActivity.java

+++ app/src/main/java/org/t_robop/y_ogawara/tancolle/SQLiteOpenHelper.java
@@ -22,7 +22,7 @@ class MySQLiteOpenHelper extends SQLiteOpenHelper {
     //ふりがな
     String kana;
     //誕生日
-    long birthday;
+    int birthday;
     //カテゴリ
     String category;
     //twitterのID
